+++ catalogue_graph/src/adapters/sources/ebsco/steps/loader.py
@@ -13,13 +13,13 @@ import pyarrow as pa
 import structlog
 from pydantic import BaseModel, ConfigDict
 
-from adapters.ebsco import helpers
-from adapters.ebsco.marcxml_loader import MarcXmlFileLoader
-from adapters.ebsco.models.step_events import (
+from adapters.sources.ebsco import helpers
+from adapters.sources.ebsco.marcxml_loader import MarcXmlFileLoader
+from adapters.sources.ebsco.models.step_events import (
     EbscoAdapterLoaderEvent,
     LoaderResponse,
 )
-from adapters.ebsco.reporting import EbscoLoaderReport
+from adapters.sources.ebsco.reporting import EbscoLoaderReport
 from adapters.utils.adapter_store import AdapterStore
 from adapters.utils.schemata import ADAPTER_STORE_ARROW_SCHEMA
 from utils.logger import ExecutionContext, get_trace_id, setup_logging
