+++ catalogue_graph/src/adapters/sources/oai_pmh/axiell/clients.py
@@ -7,8 +7,8 @@ from functools import cache
 import httpx
 from oai_pmh_client.client import OAIClient
 
-from adapters.oai_pmh.axiell import config
-from adapters.oai_pmh.http_client import OAIPMHHTTPClient
+from adapters.sources.oai_pmh.axiell import config
+from adapters.sources.oai_pmh.http_client import OAIPMHHTTPClient
 from utils.aws import get_ssm_parameter
 
 # Axiell uses a custom "Token" header for authentication
