+++ catalogue_graph/src/adapters/sources/oai_pmh/axiell/runtime.py
@@ -3,17 +3,17 @@ from __future__ import annotations
 import httpx
 from oai_pmh_client.client import OAIClient
 
-from adapters.oai_pmh.axiell.clients import (
+from adapters.sources.oai_pmh.axiell.clients import (
     _oai_endpoint,
 )
-from adapters.oai_pmh.axiell.clients import (
+from adapters.sources.oai_pmh.axiell.clients import (
     build_http_client as _build_axiell_http_client,
 )
-from adapters.oai_pmh.axiell.clients import (
+from adapters.sources.oai_pmh.axiell.clients import (
     build_oai_client as _build_axiell_oai_client,
 )
-from adapters.oai_pmh.axiell.config import AXIELL_ADAPTER_CONFIG
-from adapters.oai_pmh.runtime import OAIPMHAdapterConfig, OAIPMHRuntimeConfig
+from adapters.sources.oai_pmh.axiell.config import AXIELL_ADAPTER_CONFIG
+from adapters.sources.oai_pmh.runtime import OAIPMHAdapterConfig, OAIPMHRuntimeConfig
 from adapters.utils.iceberg import IcebergTable, get_local_table, get_rest_api_table
 
 
@@ -52,7 +52,7 @@ class AxiellRuntimeConfig(OAIPMHRuntimeConfig):
         create_if_not_exists: bool = True,
     ) -> IcebergTable:
         """Build the Iceberg table for the Axiell reconciler store."""
-        from adapters.oai_pmh.axiell.config import (
+        from adapters.sources.oai_pmh.axiell.config import (
             RECONCILER_LOCAL_CONFIG,
             RECONCILER_REST_API_CONFIG,
         )
