+++ catalogue_graph/src/adapters/sources/oai_pmh/registry.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Literal
 
-from adapters.oai_pmh.runtime import OAIPMHRuntimeConfig
+from adapters.sources.oai_pmh.runtime import OAIPMHRuntimeConfig
 
 AdapterType = Literal["axiell", "folio"]
 
@@ -14,8 +14,8 @@ _CONFIGS: dict[str, OAIPMHRuntimeConfig] | None = None
 def _load_configs() -> dict[str, OAIPMHRuntimeConfig]:
     # Lazy imports to avoid heavyweight module-level side effects
     # (SSM lookups, HTTP client construction) when not needed.
-    from adapters.oai_pmh.axiell.runtime import AXIELL_CONFIG
-    from adapters.oai_pmh.folio.runtime import FOLIO_CONFIG
+    from adapters.sources.oai_pmh.axiell.runtime import AXIELL_CONFIG
+    from adapters.sources.oai_pmh.folio.runtime import FOLIO_CONFIG
 
     return {
         "axiell": AXIELL_CONFIG,
