+++ catalogue_graph/tests/adapters/oai_pmh/conftest.py
@@ -18,10 +18,10 @@ from typing import TYPE_CHECKING
 import pytest
 from oai_pmh_client.client import OAIClient
 
-from adapters.oai_pmh.axiell.runtime import AXIELL_CONFIG
-from adapters.oai_pmh.folio.runtime import FOLIO_CONFIG
-from adapters.oai_pmh.runtime import OAIPMHRuntimeConfig
-from adapters.oai_pmh.steps.loader import LoaderRuntime
+from adapters.sources.oai_pmh.axiell.runtime import AXIELL_CONFIG
+from adapters.sources.oai_pmh.folio.runtime import FOLIO_CONFIG
+from adapters.sources.oai_pmh.runtime import OAIPMHRuntimeConfig
+from adapters.sources.oai_pmh.steps.loader import LoaderRuntime
 from adapters.utils.adapter_store import AdapterStore
 from adapters.utils.window_generator import WindowGenerator
 from adapters.utils.window_store import WindowStore
