+++ catalogue_graph/tests/adapters/oai_pmh/test_record_writer.py
@@ -7,7 +7,7 @@ import pyarrow as pa
 from lxml import etree
 from oai_pmh_client.models import Header, Record
 
-from adapters.oai_pmh.record_writer import WindowRecordWriter
+from adapters.sources.oai_pmh.record_writer import WindowRecordWriter
 from adapters.utils.adapter_store import AdapterStore
 from adapters.utils.schemata import ADAPTER_STORE_ARROW_SCHEMA
